修改

+++ java/src/main/java/toolgood/algorithm2/litJson/IJsonWrapper.java
@@ -1,27 +1,20 @@
 package toolgood.algorithm2.litJson;
 
+import java.math.BigDecimal;
+
 public interface IJsonWrapper {
-
+    
     boolean IsArray();
-
     boolean IsBoolean();
-
     boolean IsDouble();
-
     boolean IsObject();
-
     boolean IsString();
-
     boolean IsNull();
 
     void SetBoolean(boolean val);
-
-    void SetDouble(double val);
-
+    void SetDouble(BigDecimal val);
     void SetJsonType(JsonType type);
-
     void SetString(String val);
-
     void SetNull();
 
     void Add(IJsonWrapper val);
